Factor out event persistence
