+++ tests/Functional/Repository/EventRepositoryTest.php
@@ -7,10 +7,10 @@ namespace App\Tests\Functional\Repository;
 use App\Entity\Event;
 use App\Entity\Job;
 use App\Entity\Reference;
-use App\EntityFactory\ReferenceFactory;
 use App\Repository\EventRepository;
 use App\Repository\JobRepository;
 use App\Repository\ReferenceRepository;
+use App\Tests\Services\EventFactory;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use webignition\ObjectReflector\ObjectReflector;
@@ -23,7 +23,7 @@ class EventRepositoryTest extends WebTestCase
 
     private EventRepository $eventRepository;
     private JobRepository $jobRepository;
-    private ReferenceFactory $referenceFactory;
+    private EventFactory $eventFactory;
 
     protected function setUp(): void
     {
@@ -33,9 +33,9 @@ class EventRepositoryTest extends WebTestCase
         \assert($repository instanceof EventRepository);
         $this->eventRepository = $repository;
 
-        $referenceFactory = self::getContainer()->get(ReferenceFactory::class);
-        \assert($referenceFactory instanceof ReferenceFactory);
-        $this->referenceFactory = $referenceFactory;
+        $eventFactory = self::getContainer()->get(EventFactory::class);
+        \assert($eventFactory instanceof EventFactory);
+        $this->eventFactory = $eventFactory;
 
         $entityManager = self::getContainer()->get(EntityManagerInterface::class);
         \assert($entityManager instanceof EntityManagerInterface);
@@ -73,7 +73,7 @@ class EventRepositoryTest extends WebTestCase
     public function testFindByTypeScope(array $events, string $jobLabel, string $scope, array $expectedEventIds): void
     {
         foreach ($events as $event) {
-            $this->persistEvent($event);
+            $this->eventFactory->persist($event);
         }
 
         $job = $this->jobRepository->findOneBy(['label' => $jobLabel]);
@@ -187,43 +187,5 @@ class EventRepositoryTest extends WebTestCase
                 'expectedEventIds' => ['eventId1', 'eventId4'],
             ],
         ];
-    }
-
-    private function persistEvent(Event $event): void
-    {
-        $eventReference = ObjectReflector::getProperty($event, 'reference');
-        \assert($eventReference instanceof Reference);
-
-        $referenceLabel = ObjectReflector::getProperty($eventReference, 'label');
-        \assert(is_string($referenceLabel));
-        \assert('' !== $referenceLabel);
-
-        $referenceReference = ObjectReflector::getProperty($eventReference, 'reference');
-        \assert(is_string($referenceReference));
-        \assert('' !== $referenceReference);
-
-        $referenceEntity = $this->referenceFactory->create($referenceLabel, $referenceReference);
-
-        $event = $this->createEventWithReference($event, $referenceEntity);
-
-        $this->eventRepository->add($event);
-    }
-
-    private function createEventWithReference(Event $event, Reference $reference): Event
-    {
-        $reflectionClass = new \ReflectionClass($event);
-        $reflectionEvent = $reflectionClass->newInstanceWithoutConstructor();
-        \assert($reflectionEvent instanceof Event);
-
-        $referenceProperty = $reflectionClass->getProperty('reference');
-        $referenceProperty->setValue($reflectionEvent, $reference);
-
-        $propertyNames = ['id', 'sequenceNumber', 'job', 'type', 'body', 'relatedReferences'];
-        foreach ($propertyNames as $propertyName) {
-            $property = $reflectionClass->getProperty($propertyName);
-            $property->setValue($reflectionEvent, ObjectReflector::getProperty($event, $propertyName));
-        }
-
-        return $reflectionEvent;
     }
 }
